pyproject.toml
[project]
name = "fastapi-testing"
version = "0.1.0"
description = "A simple follow along of fastapi's getting started testing tutorial"
authors = [{ name = "tomit4", email = "mosssap@gmail.com" }]
dependencies = [
  "fastapi>=0.112.1",
  "uvicorn>=0.30.6",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["isort>=5.13.2", "black>=24.8.0", "pytest>=8.3.2"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_testing"]

[project.scripts]
dev = "fastapi_testing.main:main"
